requirements.txt
fastapi>=0.104.1,<0.125.0
uvicorn[standard]>=0.24.0,<0.34.0
python-dotenv>=1.0.0,<1.1.0
passlib==1.7.4
bcrypt==4.0.1
python-jose>=3.3.0,<3.5.0
email-validator>=2.0.0,<2.4.0
jsonpath-ng>=1.5.0,<1.9.0

# Tortoise-ORM 相关依赖
tortoise-orm[asyncpg,mysql,sqlite]>=0.20.0,<0.24.0
aiomysql>=0.2.0,<0.3.0
aiohttp>=3.9.0,<3.11.0